// src/services/api/generated/models/board-status-enum.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Test Management API
 * API for managing boards, tests, and execution.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * * `CONNECTED` - Connected * `DISCONNECTED` - Disconnected * `OFFLINE` - Offline * `ERROR` - Error
 */

export const BoardStatusEnum = {
    CONNECTED: 'CONNECTED',
    DISCONNECTED: 'DISCONNECTED',
    OFFLINE: 'OFFLINE',
    ERROR: 'ERROR'
} as const;

export type BoardStatusEnum = typeof BoardStatusEnum[keyof typeof BoardStatusEnum];